feat: User is added after submitting the Register form. No input validation yet
