+++ src/main/resources/templates/register.html
@@ -8,11 +8,11 @@
 	<body>
 		<div class="auth--fragment">
 				<h2>Register Here</h2>
-				<form class="auth--form">
-					<input name="register-user" placeholder="User Name"/>
-					<input name="register-email" placeholder="Email" type="email"/>
-					<input name="register-password" placeholder="Password" type="password"/>
-					<button>Register</button>
+				<form class="auth--form"action="/register" method="post">
+					<input type="text" id="name" name="name" placeholder="User Name"/>
+					<input type="email" id="email" name="email" placeholder="Email"/>
+					<input type="password" id="password" name="password" placeholder="Password"/>
+					<button type="submit">Register</button>
 				</form>
 				<p class="auth--redirect">Already have an account? <a href="/login">Log In</a></p>
 		</div>
